Implement communication between game and sidebar
Currently, this only shows a list of games in the lobby (including the ones you're not meant to see).

This also introduces a crash when changing the game's region, but reloading the game causes the region change to be applied so I'm not blocking this commit on that for now.